Make the startup dictionary demo independent of the user's dictionary
The gate asserted that the correction output contained cuBLAS and cuDNN, which
only holds if the user has already taught those terms. A fresh install has an
empty dictionary, so the loader failed at 94 percent on a clean Mac with a demo
output mismatch and the app never became usable. Give the demo its own two-entry
fixture dictionary: it now proves the retrieval and correction path works
anywhere, which is what a startup self-test should assert.

Verified end to end in a clean environment with no dictionary, no settings, no
Homebrew on PATH, no ffmpeg, and no FluidVoice: ready in 6.7s with all three
startup demos passing.

+++ sidecar/asr_server.py
@@ -73,7 +73,9 @@ def transcribe_file(model, path: str):
         return model.transcribe(path)
     from parakeet_mlx.audio import get_logmel
 
-    mel = get_logmel(mx.array(samples, dtype=mx.bfloat16), model.preprocessor_config)
+    # float32, matching what parakeet-mlx's own loader hands the preprocessor.
+    # bfloat16 changes the FFT layout and the mel matmul fails on shape.
+    mel = get_logmel(mx.array(samples, dtype=mx.float32), model.preprocessor_config)
     return model.generate(mel)[0]
